pyproject.toml
[project]
name = "normalization-service"
version = "0.1.0"
package-mode = false
description = "A collection of API-enabled normalization services"
authors = [
    {name = "devinhelgeson",email = "devinhelgeson@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "uvicorn (>=0.34.3,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "torch (==2.2.2)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "numpy (>=1.26,<2.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "datasets (>=4.0.0,<5.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "accelerate (>=1.9.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"